feat : tokenGenerateHelper

+++ Whatnow-Api/src/main/kotlin/com/depromeet/whatnow/config/jwt/JwtTokenHelper.kt
@@ -22,10 +22,10 @@ class JwtTokenHelper(
     private val jwtSecretKey: String,
 
     @Value("\${auth.jwt.access-exp}")
-    val accessExpireIn: Int,
+    val accessExpireIn: Long,
 
     @Value("\${auth.jwt.refresh-exp}")
-    val refreshExpireIn: Int,
+    val refreshExpireIn: Long,
 ) {
     fun getJws(token: String): Jws<Claims> {
         return JwsParseExecuter {
